Moved data labels in core series class.
