--- ts/Core/Series/DataLabels.ts
@@ -1,903 +1,0 @@
-/* *
- *
- *  (c) 2010-2021 Torstein Honsi
- *
- *  License: www.highcharts.com/license
- *
- *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
- *
- * */
-
-'use strict';
-
-/* *
- *
- *  Imports
- *
- * */
-
-import type AlignObject from '../Renderer/AlignObject';
-import type BBoxObject from '../Renderer/BBoxObject';
-import type ColorString from '../Color/ColorString';
-import type ColumnPoint from '../../Series/Column/ColumnPoint';
-import type DataLabelOptions from './DataLabelOptions';
-import type PiePoint from '../../Series/Pie/PiePoint';
-import type PieSeries from '../../Series/Pie/PieSeries';
-import type Point from './Point';
-import type SVGAttributes from '../Renderer/SVG/SVGAttributes';
-import type SVGElement from '../Renderer/SVG/SVGElement';
-import type SVGLabel from '../Renderer/SVG/SVGLabel';
-
-import A from '../Animation/AnimationUtilities.js';
-const { getDeferredAnimation } = A;
-import F from '../FormatUtilities.js';
-const { format } = F;
-import Palette from '../Color/Palette.js';
-import R from '../Renderer/RendererUtilities.js';
-import Series from '../Series/Series.js';
-import SeriesRegistry from './SeriesRegistry.js';
-const { seriesTypes } = SeriesRegistry;
-import U from '../Utilities.js';
-const {
-    defined,
-    extend,
-    fireEvent,
-    isArray,
-    merge,
-    objectEach,
-    pick,
-    splat
-} = U;
-
-/* *
- *
- *  Declarations
- *
- * */
-
-declare module './PointLike' {
-    interface PointLike {
-        bottom?: number;
-        connector?: SVGElement;
-        connectors?: Array<SVGElement>;
-        contrastColor?: ColorString;
-        dataLabel?: SVGLabel;
-        dataLabelOnNull?: boolean;
-        dataLabelPath?: SVGElement;
-        dataLabels?: Array<SVGElement>;
-        distributeBox?: R.BoxObject;
-        dlBox?: BBoxObject;
-        dlOptions?: DataLabelOptions;
-        graphic?: SVGElement;
-        /** @deprecated */
-        positionIndex?: unknown;
-        top?: number;
-        getDataLabelPath(dataLabel: SVGElement): SVGElement;
-    }
-}
-
-declare module './PointOptions' {
-    interface PointOptions {
-        dataLabels?: (DataLabelOptions|Array<DataLabelOptions>);
-        labelrank?: number;
-    }
-}
-
-declare module './SeriesLike' {
-    interface SeriesLike {
-        _hasPointLabels?: boolean;
-        /** @deprecated */
-        dataLabelsGroup?: SVGElement;
-        alignDataLabel(
-            point: Point,
-            dataLabel: SVGElement,
-            options: DataLabelOptions,
-            alignTo: BBoxObject,
-            isNew?: boolean
-        ): void;
-        drawDataLabels(): void;
-        justifyDataLabel(
-            dataLabel: SVGElement,
-            options: DataLabelOptions,
-            alignAttr: SVGAttributes,
-            bBox: BBoxObject,
-            alignTo?: BBoxObject,
-            isNew?: boolean
-        ): (boolean|undefined);
-        placeDataLabels?(): void;
-        setDataLabelStartPos(
-            point: ColumnPoint,
-            dataLabel: SVGElement,
-            isNew: boolean|undefined,
-            isInside: boolean,
-            alignOptions: AlignObject
-        ): void;
-        verifyDataLabelOverflow?(overflow: Array<number>): boolean;
-    }
-}
-
-declare module './SeriesOptions' {
-    interface SeriesOptions {
-        dataLabels?: (DataLabelOptions|Array<DataLabelOptions>);
-    }
-}
-
-/**
- * Callback JavaScript function to format the data label as a string. Note that
- * if a `format` is defined, the format takes precedence and the formatter is
- * ignored.
- *
- * @callback Highcharts.DataLabelsFormatterCallbackFunction
- *
- * @param {Highcharts.PointLabelObject} this
- * Data label context to format
- *
- * @param {Highcharts.DataLabelsOptions} options
- * [API options](/highcharts/plotOptions.series.dataLabels) of the data label
- *
- * @return {number|string|null|undefined}
- * Formatted data label text
- */
-
-/**
- * Values for handling data labels that flow outside the plot area.
- *
- * @typedef {"allow"|"justify"} Highcharts.DataLabelsOverflowValue
- */
-
-''; // detach doclets above
-
-/* eslint-disable valid-jsdoc */
-
-/**
- * Draw the data labels
- *
- * @private
- * @function Highcharts.Series#drawDataLabels
- * @return {void}
- * @fires Highcharts.Series#event:afterDrawDataLabels
- */
-Series.prototype.drawDataLabels = function (): void {
-    let series = this,
-        chart = series.chart,
-        seriesOptions = series.options,
-        seriesDlOptions = seriesOptions.dataLabels,
-        points = series.points,
-        pointOptions,
-        hasRendered = series.hasRendered || 0,
-        dataLabelsGroup: SVGElement,
-        dataLabelAnim = (seriesDlOptions as any).animation,
-        animationConfig = (seriesDlOptions as any).defer ?
-            getDeferredAnimation(chart, dataLabelAnim, series) :
-            { defer: 0, duration: 0 },
-        renderer = chart.renderer;
-
-    /**
-     * Handle the dataLabels.filter option.
-     * @private
-     */
-    function applyFilter(
-        point: Point,
-        options: DataLabelOptions
-    ): boolean {
-        let filter = options.filter,
-            op,
-            prop,
-            val;
-
-        if (filter) {
-            op = filter.operator;
-            prop = (point as any)[filter.property];
-            val = filter.value;
-            if (
-                (op === '>' && prop > (val as any)) ||
-                (op === '<' && prop < (val as any)) ||
-                (op === '>=' && prop >= (val as any)) ||
-                (op === '<=' && prop <= (val as any)) ||
-                (op === '==' && prop == val) || // eslint-disable-line eqeqeq
-                (op === '===' && prop === val)
-            ) {
-                return true;
-            }
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * Merge two objects that can be arrays. If one of them is an array, the
-     * other is merged into each element. If both are arrays, each element is
-     * merged by index. If neither are arrays, we use normal merge.
-     * @private
-     */
-    function mergeArrays(
-        one: (DataLabelOptions|Array<DataLabelOptions>),
-        two: (DataLabelOptions|Array<DataLabelOptions>)
-    ): (DataLabelOptions|Array<DataLabelOptions>) {
-        let res: (DataLabelOptions|Array<DataLabelOptions>) = [],
-            i: number;
-
-        if (isArray(one) && !isArray(two)) {
-            res = (one as any).map(
-                function (el: DataLabelOptions): DataLabelOptions {
-                    return merge(el, two);
-                }
-            );
-        } else if (isArray(two) && !isArray(one)) {
-            res = (two as any).map(
-                function (el: DataLabelOptions): DataLabelOptions {
-                    return merge(one, el);
-                }
-            );
-        } else if (!isArray(one) && !isArray(two)) {
-            res = merge(one, two);
-        } else {
-            i = Math.max((one as any).length, (two as any).length);
-            while (i--) {
-                (res as any)[i] = merge((one as any)[i], (two as any)[i]);
-            }
-        }
-        return res;
-    }
-
-
-    // Merge in plotOptions.dataLabels for series
-    seriesDlOptions = mergeArrays(
-        mergeArrays(
-            chart.options.plotOptions &&
-            chart.options.plotOptions.series &&
-            chart.options.plotOptions.series.dataLabels as any,
-            chart.options.plotOptions &&
-            chart.options.plotOptions[series.type] &&
-            (chart.options.plotOptions[series.type] as any).dataLabels as any
-        ),
-        seriesDlOptions as any
-    );
-
-    fireEvent(this, 'drawDataLabels');
-
-    if (
-        isArray(seriesDlOptions) ||
-        (seriesDlOptions as any).enabled ||
-        series._hasPointLabels
-    ) {
-
-        // Create a separate group for the data labels to avoid rotation
-        dataLabelsGroup = series.plotGroup(
-            'dataLabelsGroup',
-            'data-labels',
-            !hasRendered ? 'hidden' : 'inherit', // #5133, #10220
-            (seriesDlOptions as any).zIndex || 6
-        );
-
-        dataLabelsGroup.attr({ opacity: +hasRendered }); // #3300
-        if (!hasRendered) {
-            const group = series.dataLabelsGroup;
-            if (group) {
-                if (series.visible) { // #2597, #3023, #3024
-                    dataLabelsGroup.show(true);
-                }
-                (group[
-                    seriesOptions.animation ? 'animate' : 'attr'
-                ] as any)(
-                    { opacity: 1 },
-                    animationConfig
-                );
-            }
-        }
-
-        // Make the labels for each point
-        points.forEach(function (point): void {
-
-            // Merge in series options for the point.
-            // @note dataLabelAttribs (like pointAttribs) would eradicate
-            // the need for dlOptions, and simplify the section below.
-            pointOptions = splat(
-                mergeArrays(
-                    seriesDlOptions as any,
-                    (point.dlOptions as any) || // dlOptions is used in treemaps
-                        (point.options && point.options.dataLabels)
-                )
-            );
-
-            // Handle each individual data label for this point
-            pointOptions.forEach(function (
-                labelOptions: DataLabelOptions,
-                i: number
-            ): void {
-                // Options for one datalabel
-                let labelEnabled = (
-                        labelOptions.enabled &&
-                        // #2282, #4641, #7112, #10049
-                        (!point.isNull || point.dataLabelOnNull) &&
-                        applyFilter(point, labelOptions as any)
-                    ),
-                    labelConfig,
-                    formatString,
-                    labelText,
-                    style,
-                    rotation,
-                    attr: any,
-                    dataLabel: SVGLabel = point.dataLabels ? point.dataLabels[i] :
-                        point.dataLabel as any,
-                    connector = point.connectors ? point.connectors[i] :
-                        point.connector,
-                    labelDistance = pick(
-                        (labelOptions as any).distance,
-                        point.labelDistance
-                    ),
-                    isNew = !dataLabel;
-
-                if (labelEnabled) {
-                    // Create individual options structure that can be extended
-                    // without affecting others
-                    labelConfig = point.getLabelConfig();
-
-                    formatString = pick(
-                        (labelOptions as any)[point.formatPrefix + 'Format'],
-                        labelOptions.format
-                    );
-
-                    labelText = defined(formatString) ?
-                        format(formatString, labelConfig, chart) :
-                        (
-                            (labelOptions as any)[
-                                point.formatPrefix + 'Formatter'
-                            ] ||
-                            labelOptions.formatter
-                        ).call(labelConfig, labelOptions);
-
-                    style = labelOptions.style;
-                    rotation = labelOptions.rotation;
-
-                    if (!chart.styledMode) {
-                        // Determine the color
-                        (style as any).color = pick(
-                            labelOptions.color,
-                            (style as any).color,
-                            series.color,
-                            Palette.neutralColor100
-                        );
-                        // Get automated contrast color
-                        if ((style as any).color === 'contrast') {
-                            point.contrastColor = renderer.getContrast(
-                                (point.color || series.color) as any
-                            );
-                            (style as any).color = (
-                                !defined(labelDistance) &&
-                                    labelOptions.inside
-                            ) ||
-                                labelDistance < 0 ||
-                                !!seriesOptions.stacking ?
-                                point.contrastColor :
-                                Palette.neutralColor100;
-                        } else {
-                            delete point.contrastColor;
-                        }
-                        if (seriesOptions.cursor) {
-                            (style as any).cursor = seriesOptions.cursor;
-                        }
-                    }
-
-                    attr = {
-                        r: labelOptions.borderRadius || 0,
-                        rotation: rotation,
-                        padding: labelOptions.padding,
-                        zIndex: 1
-                    };
-
-                    if (!chart.styledMode) {
-                        attr.fill = labelOptions.backgroundColor;
-                        attr.stroke = labelOptions.borderColor;
-                        attr['stroke-width'] = labelOptions.borderWidth;
-                    }
-
-                    // Remove unused attributes (#947)
-                    objectEach(attr, function (val: any, name: string): void {
-                        if (typeof val === 'undefined') {
-                            delete attr[name];
-                        }
-                    });
-                }
-
-                // If the point is outside the plot area, destroy it. #678, #820
-                if (dataLabel && (!labelEnabled || !defined(labelText))) {
-                    point.dataLabel =
-                        point.dataLabel && point.dataLabel.destroy();
-                    if (point.dataLabels) {
-                        // Remove point.dataLabels if this was the last one
-                        if (point.dataLabels.length === 1) {
-                            delete point.dataLabels;
-                        } else {
-                            delete point.dataLabels[i];
-                        }
-                    }
-                    if (!i) {
-                        delete point.dataLabel;
-                    }
-                    if (connector) {
-                        point.connector = (point.connector as any).destroy();
-                        if (point.connectors) {
-                            // Remove point.connectors if this was the last one
-                            if (point.connectors.length === 1) {
-                                delete point.connectors;
-                            } else {
-                                delete point.connectors[i];
-                            }
-                        }
-                    }
-
-                // Individual labels are disabled if the are explicitly disabled
-                // in the point options, or if they fall outside the plot area.
-                } else if (labelEnabled && defined(labelText)) {
-
-                    if (!dataLabel) {
-                        // Create new label element
-                        point.dataLabels = point.dataLabels || [];
-                        dataLabel = point.dataLabels[i] = rotation ?
-
-                            // Labels don't rotate, use text element
-                            renderer.text(
-                                labelText,
-                                0,
-                                -9999,
-                                labelOptions.useHTML)
-                                .addClass('highcharts-data-label') as any :
-
-                            // We can use label
-                            renderer.label(
-                                labelText,
-                                0,
-                                -9999,
-                                labelOptions.shape as any,
-                                null as any,
-                                null as any,
-                                labelOptions.useHTML,
-                                null as any,
-                                'data-label'
-                            );
-
-                        // Store for backwards compatibility
-                        if (!i) {
-                            point.dataLabel = dataLabel;
-                        }
-
-                        dataLabel.addClass(
-                            ' highcharts-data-label-color-' + point.colorIndex +
-                            ' ' + (labelOptions.className || '') +
-                            ( // #3398
-                                labelOptions.useHTML ?
-                                    ' highcharts-tracker' :
-                                    ''
-                            )
-                        );
-                    } else {
-                        // Use old element and just update text
-                        attr.text = labelText;
-                    }
-
-                    // Store data label options for later access
-                    dataLabel.options = labelOptions;
-
-                    dataLabel.attr(attr);
-
-                    if (!chart.styledMode) {
-                        // Styles must be applied before add in order to read
-                        // text bounding box
-                        dataLabel.css(style as any).shadow(labelOptions.shadow);
-                    }
-
-                    if (!dataLabel.added) {
-                        dataLabel.add(dataLabelsGroup);
-                    }
-
-                    if (labelOptions.textPath && !labelOptions.useHTML) {
-                        dataLabel.setTextPath(
-                            (
-                                point.getDataLabelPath &&
-                                point.getDataLabelPath(dataLabel)
-                            ) || point.graphic,
-                            labelOptions.textPath
-                        );
-
-                        if (
-                            point.dataLabelPath &&
-                            !labelOptions.textPath.enabled
-                        ) {
-                            // clean the DOM
-                            point.dataLabelPath = point.dataLabelPath.destroy();
-                        }
-                    }
-
-                    // Now the data label is created and placed at 0,0, so we
-                    // need to align it
-                    series.alignDataLabel(
-                        point, dataLabel, labelOptions, null as any, isNew
-                    );
-                }
-            });
-        });
-    }
-
-    fireEvent(this, 'afterDrawDataLabels');
-};
-
-/**
- * Align each individual data label.
- *
- * @private
- * @function Highcharts.Series#alignDataLabel
- * @param {Highcharts.Point} point
- * @param {Highcharts.SVGElement} dataLabel
- * @param {Highcharts.DataLabelsOptions} options
- * @param {Highcharts.BBoxObject} alignTo
- * @param {boolean} [isNew]
- * @return {void}
- */
-Series.prototype.alignDataLabel = function (
-    point: Point,
-    dataLabel: SVGElement,
-    options: DataLabelOptions,
-    alignTo: BBoxObject,
-    isNew?: boolean
-): void {
-    let series = this,
-        chart = this.chart,
-        inverted = this.isCartesian && chart.inverted,
-        enabledDataSorting = this.enabledDataSorting,
-        plotX = pick(
-            point.dlBox && (point.dlBox as any).centerX,
-            point.plotX,
-            -9999
-        ),
-        plotY = pick(point.plotY, -9999),
-        bBox = dataLabel.getBBox(),
-        baseline,
-        rotation = options.rotation,
-        normRotation,
-        negRotation,
-        align = options.align,
-        rotCorr, // rotation correction
-        isInsidePlot = chart.isInsidePlot(
-            plotX,
-            Math.round(plotY),
-            {
-                inverted,
-                paneCoordinates: true,
-                series
-            }
-        ),
-        // Math.round for rounding errors (#2683), alignTo to allow column
-        // labels (#2700)
-        alignAttr, // the final position;
-        justify = pick(
-            options.overflow,
-            (enabledDataSorting ? 'none' : 'justify')
-        ) === 'justify',
-        visible =
-            this.visible &&
-            point.visible !== false &&
-            (
-                point.series.forceDL ||
-                (enabledDataSorting && !justify) ||
-                isInsidePlot ||
-                (
-                    // If the data label is inside the align box, it is enough
-                    // that parts of the align box is inside the plot area
-                    // (#12370). When stacking, it is always inside regardless
-                    // of the option (#15148).
-                    pick(options.inside, !!this.options.stacking) &&
-                    alignTo &&
-                    chart.isInsidePlot(
-                        plotX,
-                        inverted ?
-                            alignTo.x + 1 :
-                            alignTo.y + alignTo.height - 1,
-                        {
-                            inverted,
-                            paneCoordinates: true,
-                            series
-                        }
-                    )
-                )
-            ),
-        setStartPos = function (alignOptions: AlignObject): void {
-            if (enabledDataSorting && series.xAxis && !justify) {
-                series.setDataLabelStartPos(
-                    point as ColumnPoint,
-                    dataLabel,
-                    isNew,
-                    isInsidePlot,
-                    alignOptions
-                );
-            }
-        };
-
-    if (visible) {
-
-        baseline = chart.renderer.fontMetrics(
-            chart.styledMode ? void 0 : (options.style as any).fontSize,
-            dataLabel
-        ).b;
-
-        // The alignment box is a singular point
-        alignTo = extend({
-            x: inverted ? this.yAxis.len - plotY : plotX,
-            y: Math.round(inverted ? this.xAxis.len - plotX : plotY),
-            width: 0,
-            height: 0
-        }, alignTo);
-
-        // Add the text size for alignment calculation
-        extend<DataLabelOptions|BBoxObject>(options, {
-            width: bBox.width,
-            height: bBox.height
-        });
-
-        // Allow a hook for changing alignment in the last moment, then do the
-        // alignment
-        if (rotation) {
-            justify = false; // Not supported for rotated text
-            rotCorr = chart.renderer.rotCorr(baseline, rotation); // #3723
-            alignAttr = {
-                x: (
-                    alignTo.x +
-                    (options.x || 0) +
-                    alignTo.width / 2 +
-                    rotCorr.x
-                ),
-                y: (
-                    alignTo.y +
-                    (options.y || 0) +
-                    ({ top: 0, middle: 0.5, bottom: 1 } as any)[
-                        options.verticalAlign as any
-                    ] *
-                    alignTo.height
-                )
-            };
-            setStartPos(alignAttr); // data sorting
-            dataLabel[isNew ? 'attr' : 'animate'](alignAttr)
-                .attr({ // #3003
-                    align: align
-                });
-
-            // Compensate for the rotated label sticking out on the sides
-            normRotation = (rotation + 720) % 360;
-            negRotation = normRotation > 180 && normRotation < 360;
-
-            if (align === 'left') {
-                alignAttr.y -= negRotation ? bBox.height : 0;
-            } else if (align === 'center') {
-                alignAttr.x -= bBox.width / 2;
-                alignAttr.y -= bBox.height / 2;
-            } else if (align === 'right') {
-                alignAttr.x -= bBox.width;
-                alignAttr.y -= negRotation ? 0 : bBox.height;
-            }
-            dataLabel.placed = true;
-            dataLabel.alignAttr = alignAttr;
-
-        } else {
-            setStartPos(alignTo); // data sorting
-            dataLabel.align(options, void 0, alignTo);
-            alignAttr = dataLabel.alignAttr;
-        }
-
-        // Handle justify or crop
-        if (justify && alignTo.height >= 0) { // #8830
-            this.justifyDataLabel(
-                dataLabel,
-                options,
-                alignAttr,
-                bBox,
-                alignTo,
-                isNew
-            );
-
-        // Now check that the data label is within the plot area
-        } else if (pick(options.crop, true)) {
-            visible =
-                chart.isInsidePlot(
-                    alignAttr.x,
-                    alignAttr.y,
-                    {
-                        paneCoordinates: true,
-                        series
-                    }
-                ) &&
-                chart.isInsidePlot(
-                    alignAttr.x + bBox.width,
-                    alignAttr.y + bBox.height,
-                    {
-                        paneCoordinates: true,
-                        series
-                    }
-                );
-        }
-
-        // When we're using a shape, make it possible with a connector or an
-        // arrow pointing to thie point
-        if (options.shape && !rotation) {
-            dataLabel[isNew ? 'attr' : 'animate']({
-                anchorX: inverted ?
-                    chart.plotWidth - (point.plotY as any) :
-                    point.plotX,
-                anchorY: inverted ?
-                    chart.plotHeight - (point.plotX as any) :
-                    point.plotY
-            });
-        }
-    }
-    // To use alignAttr property in hideOverlappingLabels
-    if (isNew && enabledDataSorting) {
-        dataLabel.placed = false;
-    }
-    // Show or hide based on the final aligned position
-    if (!visible && (!enabledDataSorting || justify)) {
-        dataLabel.hide(true);
-        dataLabel.placed = false; // don't animate back in
-    }
-};
-
-/**
- * Set starting position for data label sorting animation.
- *
- * @private
- * @function Highcharts.Series#setDataLabelStartPos
- * @param {Highcharts.SVGElement} dataLabel
- * @param {Highcharts.ColumnPoint} point
- * @param {boolean | undefined} [isNew]
- * @param {boolean} [isInside]
- * @param {Highcharts.AlignObject} [alignOptions]
- *
- * @return {void}
- */
-Series.prototype.setDataLabelStartPos = function (
-    point: ColumnPoint,
-    dataLabel: SVGElement,
-    isNew: boolean,
-    isInside: boolean,
-    alignOptions: AlignObject
-): void {
-    let chart = this.chart,
-        inverted = chart.inverted,
-        xAxis = this.xAxis,
-        reversed = xAxis.reversed,
-        labelCenter = inverted ? dataLabel.height / 2 : dataLabel.width / 2,
-        pointWidth = point.pointWidth,
-        halfWidth = pointWidth ? pointWidth / 2 : 0,
-        startXPos,
-        startYPos;
-
-    startXPos = inverted ?
-        alignOptions.x :
-        (reversed ?
-            -labelCenter - halfWidth :
-            xAxis.width - labelCenter + halfWidth
-        );
-
-    startYPos = inverted ?
-        (reversed ?
-            this.yAxis.height - labelCenter + halfWidth :
-            -labelCenter - halfWidth
-        ) : alignOptions.y;
-
-    dataLabel.startXPos = startXPos;
-    dataLabel.startYPos = startYPos;
-
-    // We need to handle visibility in case of sorting point outside plot area
-    if (!isInside) {
-        dataLabel
-            .attr({ opacity: 1 })
-            .animate(
-                { opacity: 0 },
-                void 0,
-                dataLabel.hide
-            );
-
-    } else if (dataLabel.visibility === 'hidden') {
-        dataLabel.show();
-        dataLabel
-            .attr({ opacity: 0 })
-            .animate({ opacity: 1 });
-    }
-    // Save start position on first render, but do not change position
-    if (!chart.hasRendered) {
-        return;
-    }
-    // Set start position
-    if (isNew) {
-        dataLabel.attr({ x: dataLabel.startXPos, y: dataLabel.startYPos });
-    }
-
-    dataLabel.placed = true;
-};
-
-/**
- * If data labels fall partly outside the plot area, align them back in, in a
- * way that doesn't hide the point.
- *
- * @private
- * @function Highcharts.Series#justifyDataLabel
- * @param {Highcharts.SVGElement} dataLabel
- * @param {Highcharts.DataLabelsOptions} options
- * @param {Highcharts.SVGAttributes} alignAttr
- * @param {Highcharts.BBoxObject} bBox
- * @param {Highcharts.BBoxObject} [alignTo]
- * @param {boolean} [isNew]
- * @return {boolean|undefined}
- */
-Series.prototype.justifyDataLabel = function (
-    dataLabel: SVGElement,
-    options: DataLabelOptions,
-    alignAttr: SVGAttributes,
-    bBox: BBoxObject,
-    alignTo?: BBoxObject,
-    isNew?: boolean
-): (boolean|undefined) {
-    let chart = this.chart,
-        align = options.align,
-        verticalAlign = options.verticalAlign,
-        off,
-        justified,
-        padding = dataLabel.box ? 0 : (dataLabel.padding || 0);
-
-    let { x = 0, y = 0 } = options;
-
-    // Off left
-    off = (alignAttr.x || 0) + padding;
-    if (off < 0) {
-        if (align === 'right' && x >= 0) {
-            options.align = 'left';
-            options.inside = true;
-        } else {
-            x -= off;
-        }
-        justified = true;
-    }
-
-    // Off right
-    off = (alignAttr.x || 0) + bBox.width - padding;
-    if (off > chart.plotWidth) {
-        if (align === 'left' && x <= 0) {
-            options.align = 'right';
-            options.inside = true;
-        } else {
-            x += chart.plotWidth - off;
-        }
-        justified = true;
-    }
-
-    // Off top
-    off = alignAttr.y + padding;
-    if (off < 0) {
-        if (verticalAlign === 'bottom' && y >= 0) {
-            options.verticalAlign = 'top';
-            options.inside = true;
-        } else {
-            y -= off;
-        }
-        justified = true;
-    }
-
-    // Off bottom
-    off = (alignAttr.y || 0) + bBox.height - padding;
-    if (off > chart.plotHeight) {
-        if (verticalAlign === 'top' && y <= 0) {
-            options.verticalAlign = 'bottom';
-            options.inside = true;
-        } else {
-            y += chart.plotHeight - off;
-        }
-        justified = true;
-    }
-
-    if (justified) {
-        options.x = x;
-        options.y = y;
-        dataLabel.placed = !isNew;
-        dataLabel.align(options, void 0, alignTo);
-    }
-
-    return justified;
-};
